@@ -3,7 +3,7 @@ package handlers
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 
 	"github.com/gorilla/mux"
@@ -298,8 +298,8 @@ func (h *Handler) GetMesheryPatternsHandler(
 //
 // Used to get catalog patterns
 // responses:
-// 	200: mesheryPatternResponseWrapper
 //
+//	200: mesheryPatternResponseWrapper
 func (h *Handler) GetCatalogMesheryPatternsHandler(
 	rw http.ResponseWriter,
 	r *http.Request,
@@ -356,7 +356,8 @@ func (h *Handler) DeleteMesheryPatternHandler(
 //
 // Creates a local copy of a public pattern with id: id
 // responses:
-// 	200: noContentWrapper
+//
+//	200: noContentWrapper
 //
 // CloneMesheryPatternHandler clones a pattern with the given id
 func (h *Handler) CloneMesheryPatternHandler(
@@ -391,7 +392,7 @@ func (h *Handler) DeleteMultiMesheryPatternsHandler(
 	user *models.User,
 	provider models.Provider,
 ) {
-	body, err := ioutil.ReadAll(r.Body)
+	body, err := io.ReadAll(r.Body)
 	if err != nil {
 		logrus.Error(rw, "err deleting pattern, converting bytes: ", err)
 	}
