@@ -2582,7 +2582,7 @@ type CatalogFilter {
   user_id: String!
   location: Location!
   visibility: String!
-  catalog_data: Map!
+  catalog_data: Map
   created_at: String
   updated_at: String 
 }
@@ -2624,7 +2624,7 @@ type CatalogPattern {
   pattern_file: String!
   location: Location!
   visibility: String!
-  catalog_data: Map!
+  catalog_data: Map
   created_at: String
   updated_at: String
 }
@@ -4536,14 +4536,11 @@ func (ec *executionContext) _CatalogFilter_catalog_data(ctx context.Context, fie
 		return graphql.Null
 	}
 	if resTmp == nil {
-		if !graphql.HasFieldError(ctx, fc) {
-			ec.Errorf(ctx, "must not be null")
-		}
 		return graphql.Null
 	}
 	res := resTmp.(map[string]interface{})
 	fc.Result = res
-	return ec.marshalNMap2map(ctx, field.Selections, res)
+	return ec.marshalOMap2map(ctx, field.Selections, res)
 }
 
 func (ec *executionContext) fieldContext_CatalogFilter_catalog_data(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
@@ -4936,14 +4933,11 @@ func (ec *executionContext) _CatalogPattern_catalog_data(ctx context.Context, fi
 		return graphql.Null
 	}
 	if resTmp == nil {
-		if !graphql.HasFieldError(ctx, fc) {
-			ec.Errorf(ctx, "must not be null")
-		}
 		return graphql.Null
 	}
 	res := resTmp.(map[string]interface{})
 	fc.Result = res
-	return ec.marshalNMap2map(ctx, field.Selections, res)
+	return ec.marshalOMap2map(ctx, field.Selections, res)
 }
 
 func (ec *executionContext) fieldContext_CatalogPattern_catalog_data(ctx context.Context, field graphql.CollectedField) (fc *graphql.FieldContext, err error) {
@@ -16448,9 +16442,6 @@ func (ec *executionContext) _CatalogFilter(ctx context.Context, sel ast.Selectio
 
 			out.Values[i] = ec._CatalogFilter_catalog_data(ctx, field, obj)
 
-			if out.Values[i] == graphql.Null {
-				invalids++
-			}
 		case "created_at":
 
 			out.Values[i] = ec._CatalogFilter_created_at(ctx, field, obj)
@@ -16526,9 +16517,6 @@ func (ec *executionContext) _CatalogPattern(ctx context.Context, sel ast.Selecti
 
 			out.Values[i] = ec._CatalogPattern_catalog_data(ctx, field, obj)
 
-			if out.Values[i] == graphql.Null {
-				invalids++
-			}
 		case "created_at":
 
 			out.Values[i] = ec._CatalogPattern_created_at(ctx, field, obj)
