@@ -2,6 +2,7 @@ package models
 
 import (
 	"encoding/json"
+	"fmt"
 	"strings"
 
 	"github.com/gofrs/uuid"
@@ -22,7 +23,8 @@ type MesheryFilterPage struct {
 	Filters    []*MesheryFilter `json:"filters"`
 }
 
-// GetMesheryFilters returns all of the filters
+// GetMesheryFilters returns all of the 'private' filters. Though private has no meaning here since there is only
+// one local user. We make this distinction to be consistent with the remote provider
 func (mfp *MesheryFilterPersister) GetMesheryFilters(search, order string, page, pageSize uint64) ([]byte, error) {
 	order = sanitizeOrderInput(order, []string{"created_at", "updated_at", "name"})
 
@@ -33,7 +35,7 @@ func (mfp *MesheryFilterPersister) GetMesheryFilters(search, order string, page,
 	count := int64(0)
 	filters := []*MesheryFilter{}
 
-	query := mfp.DB.Order(order)
+	query := mfp.DB.Where("visibility = 'private'").Order(order)
 
 	if search != "" {
 		like := "%" + strings.ToLower(search) + "%"
@@ -54,6 +56,49 @@ func (mfp *MesheryFilterPersister) GetMesheryFilters(search, order string, page,
 	return marshalMesheryFilterPage(mesheryFilterPage), nil
 }
 
+// GetMesheryCatalogFilters returns all of the public filters
+func (mfp *MesheryFilterPersister) GetMesheryCatalogFilters(search, order string) ([]byte, error) {
+	order = sanitizeOrderInput(order, []string{"created_at", "updated_at", "name"})
+
+	if order == "" {
+		order = "updated_at desc"
+	}
+
+	filters := []*MesheryFilter{}
+
+	query := mfp.DB.Where("visibility = 'public'").Order(order)
+
+	if search != "" {
+		like := "%" + strings.ToLower(search) + "%"
+		query = query.Where("(lower(meshery_filters.name) like ?)", like)
+	}
+
+	query.Find(&filters)
+
+	marshalledFilters, _ := json.Marshal(filters)
+	return marshalledFilters, nil
+}
+
+// CloneMesheryFilter clones meshery filter to private
+func (mfp *MesheryFilterPersister) CloneMesheryFilter(filterID string) ([]byte, error) {
+	var mesheryFilter MesheryFilter
+	filterUUID, _ := uuid.FromString(filterID)
+	err := mfp.DB.First(&mesheryFilter, filterUUID).Error
+	if err != nil || *mesheryFilter.ID == uuid.Nil {
+		return nil, fmt.Errorf("unable to get filter: %w", err)
+	}
+
+	id, err := uuid.NewV4()
+	if err != nil {
+		return nil, err
+	}
+
+	mesheryFilter.Visibility = "private"
+	mesheryFilter.ID = &id
+
+	return mfp.SaveMesheryFilter(&mesheryFilter)
+}
+
 // DeleteMesheryFilter takes in a profile id and delete it if it already exists
 func (mfp *MesheryFilterPersister) DeleteMesheryFilter(id uuid.UUID) ([]byte, error) {
 	filter := MesheryFilter{ID: &id}
@@ -63,7 +108,9 @@ func (mfp *MesheryFilterPersister) DeleteMesheryFilter(id uuid.UUID) ([]byte, er
 }
 
 func (mfp *MesheryFilterPersister) SaveMesheryFilter(filter *MesheryFilter) ([]byte, error) {
-	filter.Visibility = "private"
+	if filter.Visibility == "" {
+		filter.Visibility = "private"
+	}
 	if filter.ID == nil {
 		id, err := uuid.NewV4()
 		if err != nil {
@@ -79,8 +126,12 @@ func (mfp *MesheryFilterPersister) SaveMesheryFilter(filter *MesheryFilter) ([]b
 // SaveMesheryFilters batch inserts the given filters
 func (mfp *MesheryFilterPersister) SaveMesheryFilters(filters []MesheryFilter) ([]byte, error) {
 	finalFilters := []MesheryFilter{}
+	nilUserID := ""
 	for _, filter := range filters {
-		filter.Visibility = "private"
+		if filter.Visibility == "" {
+			filter.Visibility = "private"
+		}
+		filter.UserID = &nilUserID
 		if filter.ID == nil {
 			id, err := uuid.NewV4()
 			if err != nil {
