@@ -53,7 +53,7 @@ const ArrayFieldTitle = ({ TitleField, idSchema, title, required }) => {
 
   const id = `${idSchema.$id}__title`;
   // return <h3>{title?.charAt(0)?.toUpperCase() + title?.slice(1)}</h3>;
-  return <Typography variant="body1" style={{ fontWeight : "bold", display : "inline" }}>{title.charAt(0).toUpperCase() + title.slice(1)}</Typography>;
+  return <Typography variant="body1" style={{ fontWeight : "bold", marginLeft : ".5rem", display : "inline" }}>{title.charAt(0).toUpperCase() + title.slice(1)}</Typography>;
   // return <TitleField id={id} title={title} required={required} />;
 };
 
@@ -70,10 +70,9 @@ const ArrayFieldDescription = ({ DescriptionField, idSchema, description }) => {
 const DefaultArrayItem = (props) => {
   const btnStyle = {
     flex : 1,
-    paddingLeft : 6,
+    paddingLeft : 0,
     paddingRight : 6,
     fontWeight : "bold",
-    minWidth : 0
   };
 
   return (
@@ -123,6 +122,16 @@ const DefaultArrayItem = (props) => {
 const DefaultFixedArrayFieldTemplate = (props) => {
   return (
     <fieldset className={props.className}>
+      {props.canAdd && (
+        <Button
+          className="array-item-add"
+          onClick={props.onAddClick}
+          disabled={props.disabled || props.readonly}
+        >
+          Add
+        </Button>
+      )}
+
       <ArrayFieldTitle
         key={`array-field-title-${props.idSchema.$id}`}
         TitleField={props.TitleField}
@@ -149,15 +158,7 @@ const DefaultFixedArrayFieldTemplate = (props) => {
         })}
       </div>
 
-      {props.canAdd && (
-        <Button
-          className="array-item-add"
-          onClick={props.onAddClick}
-          disabled={props.disabled || props.readonly}
-        >
-          Add
-        </Button>
-      )}
+
     </fieldset>
   );
 };
