@@ -59,22 +59,6 @@ function handleExceptionalFields(schema) {
   return returnedType || false;
 }
 
-function getDefaults(types) {
-  if (types.includes('null')) {
-    return null
-  }
-  if (types.includes('string')) {
-    return ''
-  }
-  if (types.includes('integer')) {
-    return 0
-  } if (types.includes('array')) {
-    return []
-  }
-
-  return ''
-}
-
 function addTitleToPropertiesJSONSchema(jsonSchema) {
   const newProperties = jsonSchema?.properties
 
@@ -82,36 +66,14 @@ function addTitleToPropertiesJSONSchema(jsonSchema) {
   if (newProperties && typeof newProperties === 'object') {
     Object.keys(newProperties).map(key => {
       if (Object.prototype.hasOwnProperty.call(newProperties, key)) {
-        let types = []
-
         const eField = handleExceptionalFields(newProperties[key]); // false if it is not a exceptional field
         if (eField) {
-          types = eField;
-          newProperties[key]['type'] = types
-        } else {
-          if (!Array.isArray(newProperties[key].type) && Object.prototype.hasOwnProperty.call(newProperties[key], 'type')) {
-            types.push(newProperties[key].type)
-          } else if (newProperties[key]?.type) {
-            types.push(...newProperties[key].type)
-          } else {
-            console.log("an error occurred with schema", newProperties[key])
-            throw new Error("There is something wrong with the json schema")
-          }
+          newProperties[key]['type'] = eField
         }
-
-        const defaultValue = getDefaults(types)
 
         newProperties[key] = {
           ...newProperties[key],
-          title : key,
-          default : defaultValue,
         }
-        // if (typeof newProperties[key] === 'object' && Object.prototype.hasOwnProperty.call(newProperties[key], 'properties')){
-        //   newProperties[key] = {
-        //     ...newProperties[key],
-        //     properties : addTitleToPropertiesJSONSchema(newProperties[key])
-        //   }
-        // }
       }
 
     })
